requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
Django==2.1.5
django-cors-headers==2.4.0
django-environ==0.4.5
django-tenants==2.1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
gunicorn==19.9.0
environ==1.0
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.13.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
simplegeneric==0.8.1
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
